Removed SignalChecker
It doesn't work anymore. Signal handles should be used instead for
signal handling. The idea of SignalChecker may return if the behavior
changes inside libuv, but it will not be in its current form.
Anywayit's not a big deal now since a way for proper handling of signals
is provided.