Added a new SignalChecker implementation
The eventloop does not iterate when a signal is received, so the prepare
handle trick does not work anymore. Also, polling for signals every loop
iteration doesn't sound too efficient.

Here is the new approach: use a uv_poll_t handle, and when a read
occurs, process Python signals. This is meant to be used together with
Python's signal.set_wakeup_fd. Check the example in the examples/
directory.