Keep handles alive when they are started
All handles that have start / stop methods get their refcount increased
when start is called (just once, regardless of the number of times start
was called) and decreased when stop is called.

This means that a timer can be started without holding a reference to it
(for example) and the callback will be called. Before this change the
timer would be deallocated and closed instead.