process: make Process.spawn a class method
proc = Process(loop)
proc.spawn(args, ...)

becomes:

proc = Process.spawn(loop, args, ...)

There is a slight anymmetry in uv_process_t handles: there is no
uv_process_init function, it's internal and called by uv_spawn at the
begining. So, we cannot initialize our handle in Process.__init__, only
when spawn is called, which means that a workaround was required not to
call uv_close in case a process object was initialized (in Python) but
never spawned.

This should make the interaction simpler.