pipe: add abstract namespace support
The abstract namespace is a Linux-only feature. Pipe names in the
abstract namespace start with a null byte and do not correspond to a
path in the file system.

The libuv API unfortunately makes it difficult to support these, because
both uv_pipe_bind() and uv_pipe_connect() expect a c-style null
terminate string as the pipe name. The solution is to create an AF_UNIX
socket ourselves, call bind() or connect(), and then uv_pipe_open() the
file descriptor.

The abstract namespace is used for accessing D-BUS on Linux, for
example.